+++ src/datatypes/pipeline.ts
@@ -2,6 +2,7 @@ import type { components } from "@/schema/api.js";
 
 type Schemas = components["schemas"];
 
+// Pipeline resources.
 export type Pipeline = Schemas["Pipeline"];
 export type CreatePipeline = Schemas["CreatePipeline"];
 export type UpdatePipeline = Schemas["UpdatePipeline"];